requirements.txt
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
fake-useragent==1.5.1
Faker==30.1.0
filelock==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
packaging==24.1
parsel==1.9.1
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
scrapeops_scrapy_proxy_sdk==1.0
Scrapy==2.11.2
scrapy-fake-useragent==1.4.4
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
text-unidecode==1.3
tldextract==5.1.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.0.3
